+++ rpmbuild/build.sh
@@ -56,6 +56,7 @@ Release:    1
 Summary:    ${APP_NAME}
 License:    DoD
 Requires:   httpd
+BuildArch:  noarch
 
 %description
 
@@ -73,16 +74,18 @@ Requires:   httpd
 # of the target system. This is executed on the rpmbuild machine.
 %install
 mkdir -p %{buildroot}${APP_INSTALL_DIR}
-mkdir -p %{buildroot}${APP_INSTALL_DIR}/es-home/
 mkdir -p %{buildroot}${APP_INSTALL_DIR}/httpd
-mkdir -p %{buildroot}${APP_INSTALL_DIR}/httpd/conf
-cp -a "${SCRIPT_DIR}/../es-home/build/." "%{buildroot}${APP_INSTALL_DIR}/es-home"
-cp -a "${SCRIPT_DIR}/files/." "%{buildroot}${APP_INSTALL_DIR}/httpd/conf/"
+mkdir -p %{buildroot}${APP_INSTALL_DIR}/es
+
+cp -a "${SCRIPT_DIR}/../es-home/build/." "%{buildroot}${APP_INSTALL_DIR}/es/ui"
+cp -a "${SCRIPT_DIR}/../es-common-extensions/build/." "%{buildroot}${APP_INSTALL_DIR}/es/common"
+cp -a "${SCRIPT_DIR}/files/." "%{buildroot}${APP_INSTALL_DIR}/httpd/"
 # files: List the absolute path of files to install/uninstall to the target system.
 # These are expected to also exist as root-relative under the %{buildroot}
 # directory of the buildrpm machine.
 %files
-%attr(0644, ${SCAFFOLD_USER}, ${SCAFFOLD_USER}) ${APP_INSTALL_DIR}
+%attr(0755, ${SCAFFOLD_USER}, ${SCAFFOLD_USER}) ${APP_INSTALL_DIR}
+%attr(0755, ${SCAFFOLD_USER}, ${SCAFFOLD_USER}) ${APP_INSTALL_DIR}/httpd
 
 # pre: Scripts to execute before install files to the target system
 %pre
@@ -90,9 +93,11 @@ mkdir -p "${APP_INSTALL_DIR}"
 
 # pre: Scripts to execute after installing files to the target system
 %post
-cp -r "${APP_INSTALL_DIR}/es-home" "/var/www/html/ui"
+rm -rf  "/var/www/html/es"
+mkdir -p %{buildroot}${APP_INSTALL_DIR}/httpd
+cp -r "${APP_INSTALL_DIR}/es" "/var/www/html/es"
 #copy conf.files
-cp -rfp "${APP_INSTALL_DIR}/httpd/conf/scaffold.conf" "/etc/httpd/conf.d"
+cp -rfp "${APP_INSTALL_DIR}/httpd/scaffold.conf" "/etc/httpd/conf.d"
 systemctl reload httpd 
 # preun: Scripts to execute before uninstalling files from the target system
 %preun
@@ -101,7 +106,7 @@ systemctl reload httpd
 %postun
 rm /etc/httpd/conf.d/scaffold.conf
 rm -rf ${APP_INSTALL_DIR}
-rm -rf /var/www/html/ui
+rm -rf /var/www/html/es
 systemctl reload httpd
 __EOF__
 
@@ -114,7 +119,7 @@ cd "${SCRIPT_DIR}"
 rpmbuild -v -bb --define "_topdir ${SCRIPT_DIR}" "SPECS/${APP_NAME}-${APP_VERSION_RPM}.spec"
 
 echo ""
-echo "See: ${SCRIPT_DIR}/RPMS/x86_64/${APP_NAME}-${APP_VERSION_RPM}-1.x86_64.rpm"
+echo "See: ${SCRIPT_DIR}/RPMS/noarch/${APP_NAME}-${APP_VERSION_RPM}-1.noarch.rpm"
 echo ""
 echo "[[ SUCCESS ]]"
 echo ""